restructured ticket store to include entire search response
