+++ bundle/packages/vTeams/src/redux/modules/tickets.js
@@ -1,7 +1,16 @@
-export const tickets = (state = [], action) => {
+export const tickets = (
+  state = {
+    submissions: [],
+    messages: [],
+    nextPageToken: null,
+    count: null,
+    countPageToken: null,
+  },
+  action,
+) => {
   switch (action.type) {
     case 'SET_TICKETS':
-      return [...action.payload];
+      return { ...action.payload };
     default:
       return state;
   }
